+++ R/run_simulation.R
@@ -9,6 +9,11 @@
 #' each condition
 #' @param suffix optionally a string - suffix that will be added to the name
 #' of a file storing simulation results (that will be saved to the disk)
+#' @param iter optionally a positive integer - number of MCMC iterations in
+#' model estimation
+#' @param stanPars optionally a list with additional arguments that will be
+#' passed to the Stan models' `sample` method - see [estimate_dcpo] and
+#' [estimate_claassen]
 #' @details
 #' See [run_iteration]
 #' @returns (invisibly) a list of four data frames:
@@ -34,17 +39,24 @@
 #' }
 #' @export
 run_simulation <- function(conditions, coverageScheme, nIterPerCond,
-                           suffix = "") {
+                           suffix = "", iter = 1000L, stanPars = list()) {
   check_conditions(conditions)
   check_coverage_scheme(coverageScheme, conditions)
   stopifnot(is.numeric(nIterPerCond), length(nIterPerCond) == 1,
             as.integer(nIterPerCond) == nIterPerCond, nIterPerCond > 0,
-            is.character(suffix), length(suffix) == 1L, !anyNA(suffix))
+            is.character(suffix), length(suffix) == 1L, !anyNA(suffix),
+            is.numeric(iter), length(iter) == 1L, iter > 0,
+            as.integer(iter) == iter, is.list(stanPars))
   models <- prepare_stan_models()
   modelSummaries <- countryMeans <- items <- itemDistributions <- data.frame()
   for (i in seq_len(nIterPerCond)) {
     for (j in seq_len(nrow(conditions))) {
-      resultsIter <- run_iteration(models, coverageScheme, conditions[j, ])
+      cat("\n#########################################\n Simulation iteration ",
+          i, " (out of ", nIterPerCond,"),\n condition number ", j, " (out of ",
+          nrow(conditions),")\n#########################################\n\n",
+          sep = "")
+      resultsIter <- run_iteration(models, coverageScheme, conditions[j, ],
+                                   iter, stanPars)
       modelSummaries <- dplyr::bind_rows(modelSummaries,
                                          cbind(i = i,
                                                cond = j,