+++ NEWS.md
@@ -1,8 +1,10 @@
-# cntYearTrends 0.2.0 (31.05.2023)
+# cntYearTrends 0.2.0 (1.06.2023)
 
 ## New features
 
 -   Distributions of responses to the items returned as additional element of simulation results.
+-   New arguments `iter` and `stanPars` to `run_simulation` (and `run_iteration`) allowing to provide arguments that will be passed to the Stan models `sample` method.
+-   Information about current iteration is displayed by `run_simulation`.
 -   New `check_generated_data()` function allowing to check whether the assumed simulation conditions do not lead (too often) to generating of the data with empty cells in project-country-year-item response distributions, or even worse distributions with only one non-zero cell.
 
 ## Bug fixes